商家ID调整

+++ cdt-admin/platform-shop/src/main/java/com/platform/entity/FeedbackEntity.java
@@ -45,7 +45,7 @@ public class FeedbackEntity implements Serializable {
     //
 //    private Integer msgArea;
     //商户ID
-    private Integer merchant_id;
+    private Long merchant_id;
     /**
      * 设置：主键
      */
@@ -130,11 +130,11 @@ public class FeedbackEntity implements Serializable {
         this.status = status;
     }
 
-	public Integer getMerchant_id() {
+    public long getMerchant_id() {
 		return merchant_id;
 	}
 
-	public void setMerchant_id(Integer merchant_id) {
+    public void setMerchant_id(long merchant_id) {
 		this.merchant_id = merchant_id;
 	}
 }
